Docs: remove 'development/authentication' from navigation. close #1525
Now we have item with non-existing routing in navigation panel (authentication).
It is because it is present in 'development' section of mint.json, but there is no corresponding .mdx documentation file exists.
'authentication' is related to 'deployment' section and it is already there.
Having 'authentication' in 'development' section seems to be a copy-paste error.