docs: use actual command in getting started. refs #1524
Attempts to follow the instruction and run old-fashioned command (with dash) may result in error, for example this one:

```
$ docker-compose -f docker-compose.dev.yml up
ERROR: The Compose file is invalid because:
Service keep-frontend-dev has neither an image nor a build context specified. At least one must be provided.
```

or this:
```
  File "compose/service.py", line 1579, in get_container_data_volumes
KeyError: 'ContainerConfig'
```

We should use command without dash.
See for example: https://askubuntu.com/a/1508135